fix(vehicles): correct vehicleTypeTypes import path after drain-to-feat relocation
65bc935 moved vehicleTypeTypes.ts into vehicle-types/types/ but left two
cross-feature importers in data/vehicles/ pointing at the old path, breaking
`tsc -b`. `npm run check` (prettier-only) never caught it.
